🎉 Add kc auto complet and auto rename
