+++ src/lib/KcPlayerList.ts
@@ -1,0 +1,48 @@
+import { TRiotAccount } from '@/types/RiotAccountType';
+
+export const kcPlayerList: { [key: string]: TRiotAccount[] } = {
+  // KC LEC Roster
+  Canna: [
+    { riotPseudo: 'Katze', tagLine: 'myao', region: 'EUW' },
+    { riotPseudo: 'K C', tagLine: 'kcwin', region: 'EUW' },
+  ],
+  Yike: [{ riotPseudo: 'KC Yiken', tagLine: '1111', region: 'EUW' }],
+  Kyeahoo: [{ riotPseudo: 'superkimchi', tagLine: '123', region: 'EUW' }],
+  Caliste: [
+    { riotPseudo: 'KC NEXT ADKING', tagLine: 'EUW', region: 'EUW' },
+    { riotPseudo: 'I NEED SOLOQ', tagLine: 'EUW', region: 'EUW' },
+  ],
+  Busio: [
+    { riotPseudo: 'Busio JNG', tagLine: 'NA1', region: 'EUW' },
+    { riotPseudo: 'OG Mithy', tagLine: 'Moo', region: 'EUW' },
+  ],
+  // KC Blue Roster
+  Tao: [{ riotPseudo: 'xhtao', tagLine: '3100', region: 'EUW' }],
+  Yukino: [
+    { riotPseudo: 'yukino cat', tagLine: 'blue', region: 'EUW' },
+    { riotPseudo: 'yukino cat', tagLine: 'cat', region: 'EUW' },
+  ],
+  Kamiloo: [
+    { riotPseudo: 'Limitless limits', tagLine: 'FIRE', region: 'EUW' },
+    { riotPseudo: 'TODOROKI RAICHI', tagLine: 'rank1', region: 'EUW' },
+    { riotPseudo: 'Labubu IRL', tagLine: 'macha', region: 'EUW' },
+  ],
+  Hazel: [
+    { riotPseudo: 'Antarctica', tagLine: 'S B', region: 'EUW' },
+    { riotPseudo: 'Hazel', tagLine: 'KCorp', region: 'EUW' },
+    { riotPseudo: 'Blue', tagLine: 'HZL', region: 'EUW' },
+    { riotPseudo: '114', tagLine: '1405', region: 'EUW' },
+  ],
+  Prime: [
+    { riotPseudo: 'Céleste', tagLine: '6162', region: 'EUW' },
+    { riotPseudo: 'POBC', tagLine: '6162', region: 'EUW' },
+  ],
+  // KC BS Roster
+  BAASHH: [{ riotPseudo: 'TTV BAASHH', tagLine: 'EUW', region: 'EUW' }],
+  MathisV: [{ riotPseudo: 'MathisV', tagLine: 'ARCHE', region: 'EUW' }],
+  Nsurr: [
+    { riotPseudo: 'TripleMonstre', tagLine: 'EUWFR', region: 'EUW' },
+    { riotPseudo: 'Full Drip Nsurr', tagLine: 'EUW', region: 'EUW' },
+    { riotPseudo: 'Nsurr', tagLine: 'EUWFR', region: 'EUW' },
+  ],
+};
